+++ vitest.config.ts
@@ -47,6 +47,12 @@ export default defineConfig({
     coverage: {
       provider: 'v8',
       reporter: ['text', 'html'],
+      thresholds: {
+        statements: 90,
+        branches: 75,
+        functions: 90,
+        lines: 90,
+      },
       include: ['lib/**/*.ts'],
       exclude: [
         'lib/**/*.spec.ts',